# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZeroAgent Event Venue Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  user: "user",
  owner: "owner",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role?: RegisterBodyRole;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  owner: "owner",
  user: "user",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export type VenueStatus = (typeof VenueStatus)[keyof typeof VenueStatus];

export const VenueStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Venue {
  id: number;
  ownerId: number;
  ownerName?: string;
  name: string;
  description: string;
  city: string;
  address: string;
  capacity: number;
  pricePerDay: number;
  eventTypes: string[];
  images: string[];
  status: VenueStatus;
  createdAt: string;
}

export type VenueDetail = Venue & {
  bookedDates?: string[];
};

export interface VenueListResponse {
  venues: Venue[];
  total: number;
  page: number;
  totalPages: number;
}

export interface CreateVenueBody {
  name: string;
  description: string;
  city: string;
  address: string;
  capacity: number;
  pricePerDay: number;
  eventTypes: string[];
  images?: string[];
}

export interface UpdateVenueBody {
  name?: string;
  description?: string;
  city?: string;
  address?: string;
  capacity?: number;
  pricePerDay?: number;
  eventTypes?: string[];
  images?: string[];
}

export interface PlatformStats {
  totalVenues: number;
  totalBookings: number;
  totalCities: number;
  totalUsers: number;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
  cancelled: "cancelled",
} as const;

export type BookingPaymentStatus =
  (typeof BookingPaymentStatus)[keyof typeof BookingPaymentStatus];

export const BookingPaymentStatus = {
  unpaid: "unpaid",
  paid: "paid",
} as const;

export interface Booking {
  id: number;
  venueId: number;
  venueName: string;
  userId: number;
  userName: string;
  eventDate: string;
  eventType: string;
  guestCount: number;
  status: BookingStatus;
  paymentStatus: BookingPaymentStatus;
  notes?: string;
  rejectionReason?: string;
  createdAt: string;
}

export type BookingDetail = Booking & {
  venue?: Venue;
  chatEnabled?: boolean;
};

export interface CreateBookingBody {
  venueId: number;
  eventDate: string;
  eventType: string;
  guestCount: number;
  notes?: string;
}

export interface RejectBookingBody {
  reason?: string;
}

export type UpdatePaymentBodyPaymentStatus =
  (typeof UpdatePaymentBodyPaymentStatus)[keyof typeof UpdatePaymentBodyPaymentStatus];

export const UpdatePaymentBodyPaymentStatus = {
  unpaid: "unpaid",
  paid: "paid",
} as const;

export interface UpdatePaymentBody {
  paymentStatus: UpdatePaymentBodyPaymentStatus;
}

export interface ChatMessage {
  id: number;
  bookingId: number;
  senderId: number;
  senderName: string;
  senderRole: string;
  message: string;
  createdAt: string;
}

export interface SendMessageBody {
  message: string;
}

export interface AdminStats {
  totalUsers: number;
  totalOwners: number;
  totalVenues: number;
  pendingVenues: number;
  totalBookings: number;
  pendingBookings: number;
  approvedBookings: number;
  totalRevenue: number;
  paidBookings: number;
  unpaidBookings: number;
}

export type ListVenuesParams = {
  city?: string;
  minCapacity?: number;
  maxPrice?: number;
  eventType?: string;
  page?: number;
  limit?: number;
};

export type ListBookingsParams = {
  status?: ListBookingsStatus;
};

export type ListBookingsStatus =
  (typeof ListBookingsStatus)[keyof typeof ListBookingsStatus];

export const ListBookingsStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
  cancelled: "cancelled",
} as const;