# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ZeroAgent Event Venue Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /venues:
    get:
      operationId: listVenues
      tags: [venues]
      summary: List all approved venues
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: minCapacity
          in: query
          schema:
            type: integer
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: eventType
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: List of venues
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueListResponse"

    post:
      operationId: createVenue
      tags: [venues]
      summary: Create a new venue listing (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVenueBody"
      responses:
        "201":
          description: Venue created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings (filtered by role)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, cancelled]
      responses:
        "200":
          description: Bookings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"

    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingDetail"
  /bookings/{id}/approve:
    post:
      operationId: approveBooking
      tags: [bookings]
      summary: Approve a booking request (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/reject:
    post:
      operationId: rejectBooking
      tags: [bookings]
      summary: Reject a booking request (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectBookingBody"
      responses:
        "200":
          description: Booking rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/cancel:
    post:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking (user only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}/payment:
    patch:
      operationId: updatePaymentStatus
      tags: [bookings]
      summary: Update payment status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaymentBody"
      responses:
        "200":
          description: Payment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /admin/venues/{id}/approve:
    post:
      operationId: adminApproveVenue
      tags: [admin]
      summary: Approve a venue listing (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Venue approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
  /admin/venues/{id}/reject:
    post:
      operationId: adminRejectVenue
      tags: [admin]
      summary: Reject a venue listing (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Venue rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [user, owner]
          default: user
      required: [name, email, password]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required: [user, message]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, owner, user]
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, createdAt]
    Venue:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        ownerName:
          type: string
        name:
          type: string
        description:
          type: string
        city:
          type: string
        address:
          type: string
        capacity:
          type: integer
        pricePerDay:
          type: number
        eventTypes:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
          format: date-time
      required: [id, ownerId, name, description, city, address, capacity, pricePerDay, eventTypes, images, status, createdAt]
...
    VenueListResponse:
      type: object
      properties:
        venues:
          type: array
          items:
            $ref: "#/components/schemas/Venue"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [venues, total, page, totalPages]
    CreateVenueBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        city:
          type: string
        address:
          type: string
        capacity:
          type: integer
        pricePerDay:
          type: number
        eventTypes:
          type: array
          items:
            type: string
        images:
          type: array
          items:
            type: string
      required: [name, description, city, address, capacity, pricePerDay, eventTypes]
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        venueId:
          type: integer
        venueName:
          type: string
        userId:
          type: integer
        userName:
          type: string
        eventDate:
          type: string
          format: date
        eventType:
          type: string
        guestCount:
          type: integer
        status:
          type: string
          enum: [pending, approved, rejected, cancelled]
        paymentStatus:
          type: string
          enum: [unpaid, paid]
        notes:
          type: string
        rejectionReason:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, venueId, venueName, userId, userName, eventDate, eventType, guestCount, status, paymentStatus, createdAt]
    BookingDetail:
      allOf:
        - $ref: "#/components/schemas/Booking"
        - type: object
          properties:
            venue:
              $ref: "#/components/schemas/Venue"
            chatEnabled:
              type: boolean
    CreateBookingBody:
      type: object
      properties:
        venueId:
          type: integer
        eventDate:
          type: string
          format: date
        eventType:
          type: string
        guestCount:
          type: integer
        notes:
          type: string
      required: [venueId, eventDate, eventType, guestCount]
    RejectBookingBody:
      type: object
      properties:
        reason:
          type: string
    UpdatePaymentBody:
      type: object
      properties:
        paymentStatus:
          type: string
          enum: [unpaid, paid]
      required: [paymentStatus]
...