# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZeroAgent Event Venue Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingPaymentStatus } from "./bookingPaymentStatus";
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: number;
  venueId: number;
  venueName: string;
  userId: number;
  userName: string;
  eventDate: Date;
  eventType: string;
  guestCount: number;
  status: BookingStatus;
  paymentStatus: BookingPaymentStatus;
  notes?: string;
  rejectionReason?: string;
  createdAt: Date;
}