# lib/api-zod/src/generated/types/createBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZeroAgent Event Venue Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingBody {
  venueId: number;
  eventDate: Date;
  eventType: string;
  guestCount: number;
  notes?: string;
}