# lib/api-zod/src/generated/types/registerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZeroAgent Event Venue Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { RegisterBodyRole } from "./registerBodyRole";

export interface RegisterBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role?: RegisterBodyRole;
}