# lib/api-zod/src/generated/types/venue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZeroAgent Event Venue Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { VenueStatus } from "./venueStatus";

export interface Venue {
  id: number;
  ownerId: number;
  ownerName?: string;
  name: string;
  description: string;
  city: string;
  address: string;
  capacity: number;
  pricePerDay: number;
  eventTypes: string[];
  images: string[];
  status: VenueStatus;
  createdAt: Date;
}