# lib/api-zod/src/generated/types/venueListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZeroAgent Event Venue Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Venue } from "./venue";

export interface VenueListResponse {
  venues: Venue[];
  total: number;
  page: number;
  totalPages: number;
}